# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/scooters": {
      "post": {
        "operationId": "ScootersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScooterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scooter"
                }
              }
            }
          }
        },
        "tags": [
          "scooters"
        ]
      },
      "get": {
        "operationId": "ScootersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Scooter"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "scooters"
        ]
      }
    },
    "/scooters/{id}": {
      "get": {
        "operationId": "ScootersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scooter"
                }
              }
            }
          }
        },
        "tags": [
          "scooters"
        ]
      },
      "put": {
        "operationId": "ScootersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScooterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scooter"
                }
              }
            }
          }
        },
        "tags": [
          "scooters"
        ]
      },
      "delete": {
        "operationId": "ScootersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "scooters"
        ]
      }
    },
    "/rents": {
      "get": {
        "operationId": "RentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rent"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "rents"
        ]
      },
      "post": {
        "operationId": "RentsController_rent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rent"
                }
              }
            }
          }
        },
        "tags": [
          "rents"
        ]
      }
    },
    "/rents/{id}": {
      "get": {
        "operationId": "RentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rent"
                }
              }
            }
          }
        },
        "tags": [
          "rents"
        ]
      },
      "put": {
        "operationId": "RentsController_return",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnScooterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rent"
                }
              }
            }
          }
        },
        "tags": [
          "rents"
        ]
      }
    }
  },
  "info": {
    "title": "WEMO Homework",
    "description": "A simple scooter rent service",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "scooters",
      "description": ""
    },
    {
      "name": "users",
      "description": ""
    },
    {
      "name": "rents",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "age",
          "isActive"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "age"
        ]
      },
      "CreateScooterDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string"
          },
          "typeId": {
            "type": "number"
          }
        },
        "required": [
          "plate",
          "typeId"
        ]
      },
      "Scooter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "plate": {
            "type": "string"
          },
          "typeId": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "plate",
          "typeId",
          "isActive"
        ]
      },
      "UpdateScooterDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string"
          },
          "typeId": {
            "type": "number"
          }
        }
      },
      "Rent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "scooterId": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          },
          "rentAt": {
            "format": "date-time",
            "type": "string"
          },
          "returnAt": {
            "format": "date-time",
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "scooterId",
          "userId",
          "rentAt",
          "returnAt",
          "token"
        ]
      },
      "CreateRentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "scooterId": {
            "type": "number"
          }
        },
        "required": [
          "userId",
          "scooterId"
        ]
      },
      "ReturnScooterDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "token"
        ]
      }
    }
  }
}